OpenVINO For Smart City
Application "OpenVINO for Smart City" [0], this project consists on
showcasing the advantages of the Intel’s OpenVINO toolkit. We will
develop a Near Misses case scenario, where we will detect vehicles and
pedestrians and estimate a metric of a crossroad’s dangerousness. For
that, we will use the OpenVINO toolkit and OpenCV, all written in C++.

[0] https://github.com/incluit/OpenVino-For-SmartCity/

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>